[#22]feat: classroom과 student 도메인 분리
