+++ src/main/java/soma/haeya/edupi_db/student/repository/StudentRepository.java
@@ -1,4 +1,4 @@
-package soma.haeya.edupi_db.classroom.repository;
+package soma.haeya.edupi_db.student.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import soma.haeya.edupi_db.classroom.domain.Student;
